#853 Moved test configuration instatiation from DB creator to test runner.
